ダイアログの実装をまとめよう
