+++ app/src/main/java/net/ichigotake/android/common/app/BaseDialogFragment.java
@@ -1,0 +1,23 @@
+package net.ichigotake.android.common.app;
+
+import android.app.DialogFragment;
+
+public abstract class BaseDialogFragment extends DialogFragment {
+
+    private boolean willDismiss;
+
+    @Override
+    public void onResume() {
+        super.onResume();
+        if (willDismiss) {
+            dismiss();
+        }
+    }
+
+    @Override
+    public void dismiss() {
+        super.dismiss();
+        willDismiss = true;
+    }
+
+}
